requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
anyascii==0.3.3
anyio==4.13.0
attrs==26.1.0
audioread==3.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
contourpy==1.3.3
coqpit-config==0.2.5
coqui-tts==0.27.5
coqui-tts-trainer==0.3.3
cycler==0.12.1
decorator==5.2.1
docopt==0.6.2
einops==0.8.2
filelock==3.25.2
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
inflect==7.5.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
ko-speech-tools==0.1.0
lazy-loader==0.5
librosa==0.11.0
llvmlite==0.47.0
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
monotonic-alignment-search==0.2.1
more-itertools==11.0.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
networkx==3.6.1
num2words==0.5.14
numba==0.65.0
numpy==2.4.4
packaging==26.1
pillow==12.2.0
platformdirs==4.9.6
pooch==1.9.0
propcache==0.4.1
protobuf==7.34.1
psutil==7.2.2
pycparser==3.0
pydub==0.25.1
Pygments==2.20.0
pyparsing==3.3.2
pysbd==0.3.4
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
rich==15.0.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
shellingham==1.5.4
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.4.1+cu121
torchaudio==2.4.1+cu121
tqdm==4.67.3
transformers==4.57.6
typeguard==4.5.1
typer==0.24.1
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
yarl==1.23.0
